Moved continuousintegration to pkg
